apps/ocr-service/pyproject.toml
[project]
name = "personel-ocr"
version = "0.1.0"
description = "OCR service for Personel Platform Phase 2.8 — screenshot text extraction with KVKK-compliant PII redaction"
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.34",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "pillow>=11,<12",
    "pytesseract>=0.3.10,<0.4.0",
    "paddleocr>=2.8,<3.0",
    "numpy>=2,<3",
    "python-multipart>=0.0.9,<0.0.20",
    "prometheus-client>=0.21,<0.22",
    "structlog>=24.0,<25.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<0.26",
    "httpx>=0.27,<0.29",
    "ruff>=0.7,<0.9",
    "anyio>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/personel_ocr"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["personel_ocr"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
